+++ routes/api.ts
@@ -4,5 +4,6 @@ const router = Router()
 
 // Api routes
 router.get('/convert/:from?/:to?/:amount?/:date?', ConverterController.convert)
+router.post('/multi-convert', ConverterController.multiConvert)
 
 export default router;